+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q33/Solution.java
@@ -1,11 +1,11 @@
-package com.noob.algorithm.leetcode.q033;
+package com.noob.algorithm.leetcode.q33;
 
 import java.util.Arrays;
 
 /**
- * 🟡 033.搜索旋转排序数组 - https://leetcode.cn/problems/search-in-rotated-sorted-array/description/
+ * 33.搜索旋转排序数组
  */
-public class Solution1 {
+public class Solution {
     public int search(int[] nums, int target) {
         // 区分有无旋转两种情况，通过判断nums是否完全升序来界定
         int validOrderRes = validOrder(nums);
@@ -14,8 +14,8 @@ public class Solution1 {
             return binarySearch(nums, target);
         } else {
             // 有旋转，基于轴点分别进行二分检索，返回最终检索值(copyOfRange [from,to))
-            int[] nums1 = Arrays.copyOfRange(nums, 0, validOrderRes + 1);
-            int[] nums2 = Arrays.copyOfRange(nums, validOrderRes + 1, nums.length);
+            int[] nums1 = Arrays.copyOfRange(nums, 0, validOrderRes+1);
+            int[] nums2 = Arrays.copyOfRange(nums, validOrderRes+1, nums.length);
             int search1 = binarySearch(nums1, target);
             int search2 = binarySearch(nums2, target);
             if (search1 != -1) {
@@ -34,7 +34,7 @@ public class Solution1 {
             return -1;
         }
         // 如果完全升序则返回-1，如果非完全升序则返回"轴点"（出现降序的索引位置）
-        for (int i = 0; i < nums.length - 1; i++) {
+        for (int i = 0; i < nums.length -1; i++) {
             if (nums[i] > nums[i + 1]) {
                 return i;
             }
@@ -63,5 +63,14 @@ public class Solution1 {
         }
         // 无匹配结果
         return -1;
+    }
+
+    public static void main(String[] args) {
+//        int[] nums = {4, 5, 6, 7, 0, 1, 2};
+//        int[] nums = {4, 5, 6, 7, 0, 1, 2};
+//        int[] nums = {1,3};
+        int[] nums = {3,1};
+        Solution solution = new Solution();
+        System.out.println(solution.search(nums, 1));
     }
 }
